Refactor entitlement helpers
This commit moves the entitlement helpers out of the Run Tasks integration
tests into a more generic test helper file.
