+++ run_task_integration_test.go
@@ -5,41 +5,12 @@ package tfe
 
 import (
 	"context"
-	"errors"
 	"os"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 )
-
-func getOrgEntitlements(client *Client, organizationName string) (*Entitlements, error) {
-	ctx := context.Background()
-	orgEntitlements, err := client.Organizations.ReadEntitlements(ctx, organizationName)
-	if err != nil {
-		return nil, err
-	}
-	if orgEntitlements == nil {
-		return nil, errors.New("The organization entitlements are empty.")
-	}
-	return orgEntitlements, nil
-}
-
-func hasGlobalRunTasks(client *Client, organizationName string) (bool, error) {
-	oe, err := getOrgEntitlements(client, organizationName)
-	if err != nil {
-		return false, err
-	}
-	return oe.GlobalRunTasks, nil
-}
-
-func hasPrivateRunTasks(client *Client, organizationName string) (bool, error) {
-	oe, err := getOrgEntitlements(client, organizationName)
-	if err != nil {
-		return false, err
-	}
-	return oe.PrivateRunTasks, nil
-}
 
 func TestRunTasksCreate(t *testing.T) {
 	client := testClient(t)
